moving SVG export tests BiTemporalQueryUtilTest -> SVGexportTest
